+++ run_examples.sh
@@ -17,6 +17,13 @@ examples=(
     "hash"
     "fibonacci"
     "comprehensive"
+    "for_loops"
+    "break_continue"
+    "logical_operators"
+    "string_methods"
+    "math_functions"
+    "array_methods"
+    "new_features"
 )
 
 for example in "${examples[@]}"; do